Adding a If check to avoid refactorig failures
git-svn-id: https://svn.wso2.org/repos/wso2/trunk/tools@152719 a5903396-d722-0410-b921-86c7d4935375

+++ ide/eclipse/esb/org.wso2.developerstudio.eclipse.artifact.endpoint/src/org/wso2/developerstudio/eclipse/artifact/endpoint/refactor/EndpointArtifactDeleteParticipant.java
@@ -158,12 +158,16 @@ public class EndpointArtifactDeleteParticipant extends DeleteParticipant{
 								+ originalFile.getName().substring(0,
 										originalFile.getName().length() - 3)
 								+ "esb");
-				DeleteResourceChange deleteResourceChange = new DeleteResourceChange(
-						file.getFullPath(), true, true);
-				DeleteResourceChange deleteResourceChange1 = new DeleteResourceChange(
-						file2.getFullPath(), true, true);
-				deleteChange.add(deleteResourceChange);
-				deleteChange.add(deleteResourceChange1);
+				if (file.exists()) {
+					DeleteResourceChange deleteResourceChange = new DeleteResourceChange(
+							file.getFullPath(), true, true);
+					deleteChange.add(deleteResourceChange);
+				}
+				if (file2.exists()) {
+					DeleteResourceChange deleteResourceChange1 = new DeleteResourceChange(
+							file2.getFullPath(), true, true);
+					deleteChange.add(deleteResourceChange1);
+				}
 			}
 		}
